bus refacto complete. TODO DEBUG

+++ Business/ThirdParty.Contract/Services/IPricingQueryService.cs
@@ -5,7 +5,7 @@ public interface IPricingQueryService
     Price GetPrice
     (
         //room
-        int personMaxCount, int floorNum, int floorNumMax, int hotelRank, int latitude, int longitude,
+        int personMaxCount, int floorNum, int floorNumMax, int hotelRank, double latitude, double longitude,
 
         //booking
         int personCount, DateTime arrivalDate, DateTime departureDate, string? currency, CustomerProfile? customerProfile
